Update: Changed File Name
Updated the script file name and class to be just Tower, instead of tower controller. Also made some minor changes to make improvements on the code
